+++ src/bin/main.rs
@@ -1,5 +1,15 @@
 extern crate tetris;
+use std::env::args;
 
 fn main() {
-    tetris::run()
+    let train = match args().nth(1) {
+        Some(arg) if arg == "--train" => true,
+        _  => false,
+    };
+    let dry = match args().nth(2) {
+        Some(arg) if arg == "--dry" => true,
+        _  => false,
+    };
+
+    tetris::run(train, dry);
 }
